task_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
pika==1.3.2
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary
boto3
django-cors-headers==4.3.1